+++ lib/Helper/Filter/DB/AbstractRecipeFilter.php
@@ -1,16 +1,16 @@
 <?php
 
-namespace OCA\Cookbook\Helper\Filter;
+namespace OCA\Cookbook\Helper\Filter\DB;
 
 use OCA\Cookbook\Exception\InvalidRecipeException;
 use OCP\Files\File;
 
 /**
- * An abstract filter on a recipe JSON.
+ * An abstract filter on a recipe.
  *
  * A filter should have a single purpose that is serves and implement this interface
  */
-abstract class AbstractJSONFilter {
+interface AbstractRecipeFilter {
 	/**
 	 * Filter the given recipe according to the filter class specification.
 	 *
@@ -18,22 +18,9 @@ abstract class AbstractJSONFilter {
 	 * In order to signal if the JSON file needs updating, the return value must be true if and only if the recipe was changed.
 	 *
 	 * @param array $json The recipe data as array
+	 * @param File $recipe The file containing the recipe for further processing
 	 * @return bool true, if and only if the recipe was changed
 	 * @throws InvalidRecipeException if the recipe was not correctly filtered
 	 */
-	abstract public function apply(array &$json): bool;
-
-	/**
-	 * @param string|int|float|array $value
-	 */
-	protected function setJSONValue(array &$json, string $key, string|int|float|array $value): bool {
-		if (!array_key_exists($key, $json)) {
-			$json[$key] = $value;
-			return true;
-		} elseif ($json[$key] !== $value) {
-			$json[$key] = $value;
-			return true;
-		}
-		return false;
-	}
+	public function apply(array &$json, File $recipe): bool;
 }
