+++ lib/Helper/Filter/JSON/SchemaConformityFilter.php
@@ -2,8 +2,6 @@
 
 namespace OCA\Cookbook\Helper\Filter\JSON;
 
-use OCA\Cookbook\Helper\Filter\AbstractJSONFilter;
-
 /**
  * Make sure the schema.org standard annotations are present in the JSON object.
  */
