+++ lib/Helper/Filter/Output/RecipeJSONOutputFilter.php
@@ -1,8 +1,7 @@
 <?php
 
-namespace OCA\Cookbook\Helper\Filter;
+namespace OCA\Cookbook\Helper\Filter\Output;
 
-use OCA\Cookbook\Helper\Filter\Output\EnsureNutritionPresentFilter;
 use OCP\Files\File;
 
 class RecipeJSONOutputFilter {
