+++ lib/Helper/Filter/Output/RecipeStubFilter.php
@@ -1,7 +1,8 @@
 <?php
 
-namespace OCA\Cookbook\Helper\Filter;
+namespace OCA\Cookbook\Helper\Filter\Output;
 
+use OCA\Cookbook\Helper\Filter\JSON\AbstractJSONFilter;
 use OCA\Cookbook\Helper\Filter\JSON\RecipeIdCopyFilter;
 use OCA\Cookbook\Helper\Filter\JSON\RecipeIdTypeFilter;
