+++ tests/Unit/Helper/Filter/NormalizeRecipeFileFilterTest.php
@@ -2,8 +2,8 @@
 
 namespace OCA\Cookbook\tests\Unit\Helper\Filter;
 
+use OCA\Cookbook\Helper\Filter\DB\NormalizeRecipeFileFilter;
 use OCA\Cookbook\Helper\Filter\DB\RecipeDatesFilter;
-use OCA\Cookbook\Helper\Filter\NormalizeRecipeFileFilter;
 use OCP\Files\File;
 use PHPUnit\Framework\MockObject\MockObject;
 use PHPUnit\Framework\TestCase;
